Changed variable types from float to double, 1/2 through calculateNoise method

+++ level/Fade.java
@@ -1,13 +1,13 @@
 package level;
 
 public class Fade {
-    float x;
+    double x;
     
-    public Fade(float x) {
+    public Fade(double x) {
         this.x = x;
     }
 
-    public float calculateFade() {
+    public double calculateFade() {
         return ((6 * x - 15) * x + 10) * x * x * x;
     }
 }
